+++ src/components/custom/CardMusic.tsx
@@ -3,15 +3,18 @@ import {
   CardContent,
   CardFooter,
 } from "@/components/ui/card"
-import type { CardMusicProps } from "@/interfaces/ICardMusic"
+import { motion } from "framer-motion"
 
+import type { CardMusicProps } from "@/interfaces/ICardMusic"
 
 const CardMusic = ({ url, titulo, subtitulo, width = 200, height = 200 }: CardMusicProps) => {
   return (
     <Card>
-      <CardContent>
-        <img alt="imagen" src={url} width={width} height={height} className="rounded-lg" />
-      </CardContent>
+      <motion.div whileHover={{ scale: 1.1 }} whileTap={{ scale: 0.9 }} >
+        <CardContent>
+          <img alt="imagen" src={url} width={width} height={height} className="rounded-lg" />
+        </CardContent>
+      </motion.div>
       <CardFooter className="flex flex-col gap-3 my-4 mx-auto">
         <h3 className="font-semibold">{titulo}</h3>
         <p className="font-thin">{subtitulo}</p>
