+++ src/components/custom/Main.tsx
@@ -2,11 +2,9 @@ import TabsMenu from "./TabsMenu"
 
 const Main = () => {
   return (
-    <main className="flex flex-col justify-between p-2  ml-6 w-full">
+    <main className="flex flex-col justify-between p-2 w-full">
       {/* <section>menu toggle + boton1</section> */}
       <TabsMenu />
-      <section>carousel principal</section>
-      <section>corousel secundario</section>
     </main>
   )
 }
