+++ src/lib/data.ts
@@ -1,0 +1,46 @@
+import type { CardMusicProps } from "@/interfaces/ICardMusic"
+
+const cards: CardMusicProps[] = [
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 1",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 2",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 3",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 4",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 1",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 2",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 3",
+    subtitulo: "Un subtitulo de demostracion"
+  },
+  {
+    url: "https://cdn.pixabay.com/photo/2024/04/12/23/11/nature-8692760_1280.jpg",
+    titulo: "Titulo general 4",
+    subtitulo: "Un subtitulo de demostracion"
+  }
+]
+
+export default cards
